remove duplicate stream
