+++ tests/daal4py/sycl/covariance_streaming.py
@@ -27,7 +27,6 @@ import daal4py as d4p
 from daal4py.oneapi import sycl_buffer
 
 sys.path.insert(0, "..")
-from stream import read_next
 
 try:
     from daal4py.oneapi import sycl_context
@@ -36,6 +35,24 @@ try:
         gpu_available = True
 except:
     gpu_available = False
+
+
+# a generator which reads a file in chunks
+def read_next(file, chunksize, readcsv=read_csv):
+    assert os.path.isfile(file)
+    s = 0
+    while True:
+        # if found a smaller chunk we set s to < 0 to indicate eof
+        if s < 0:
+            return
+        a = read_csv(file, s=s, n=chunksize)
+        # last chunk is usually smaller, if not,
+        # numpy will print warning in next iteration
+        if chunksize > a.shape[0]:
+            s = -1
+        else:
+            s += a.shape[0]
+        yield a
 
 
 # At this moment with sycl we are working only with numpy arrays
